Added imports, fixed function return types and got post endpoint to work

+++ src/main/java/edu/ucsb/cs156/example/documents/Earthquake/EarthquakeMetadata.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class EarthquakeMetadata {
-    private int generated;
+    private Double generated;
     private String url;
     private String title;
     private int status;
